requirements.txt
numpy

sys                  #For printing info on screen
skimage.feature.hog  #For generating hog features
matplotlib           #For generating plots
